+++ packages/firebase_admin_sdk/lib/src/app_check/app_check.dart
@@ -13,10 +13,10 @@
 // limitations under the License.
 
 import 'dart:async';
+import 'dart:convert';
 
 import 'package:googleapis/firebaseappcheck/v1.dart' as appcheck1;
 import 'package:googleapis_auth/auth_io.dart' as googleapis_auth;
-import 'package:googleapis_beta/firebaseappcheck/v1beta.dart' as appcheck1_beta;
 import 'package:meta/meta.dart';
 
 import '../app.dart';